fix(SDKChooser): camera rig not enabled correctly
When loading any non-desktop SDK (every SDK except the simulator) the
`VRSettings` were not enabled which lead to the VR camera rig objects
getting disabled in the case of SteamVR for example. This fix makes sure
to enable the `VRSettings` before loading the VR scene and additionally
disables the SDK Chooser's canvas so it's not visible while loading the
VR scene.